requirements.txt
selenium
pytesseract
requests
pyyaml